+++ Items/Ammo/Arrow/MartianArrow.cs
@@ -3,7 +3,7 @@ using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SpiritMod.Items.Ammo
+namespace SpiritMod.Items.Ammo.Arrow
 {
 	class MartianArrow : ModItem
 	{
