+++ Items/Ammo/Arrow/PoisonArrow.cs
@@ -2,7 +2,7 @@
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SpiritMod.Items.Ammo
+namespace SpiritMod.Items.Ammo.Arrow
 {
 	class PoisonArrow : ModItem
 	{
