+++ Items/Ammo/Arrow/SpectreArrow.cs
@@ -1,14 +1,14 @@
 ﻿using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SpiritMod.Items.Ammo
+namespace SpiritMod.Items.Ammo.Arrow
 {
-	class ShroomiteArrow : ModItem
+	class SpectreArrow : ModItem
 	{
 		public override void SetStaticDefaults()
 		{
-			DisplayName.SetDefault("Shroomite Arrow");
-			Tooltip.SetDefault("Flies straight and deals two ticks of damage to hit enemies!");
+			DisplayName.SetDefault("Ghast Arrow");
+			Tooltip.SetDefault("Hitting enemies occasionally releases damaging spectre bolts");
 		}
 
 		public override void SetDefaults()
@@ -20,21 +20,21 @@ namespace SpiritMod.Items.Ammo
 
 			item.maxStack = 999;
 
-			item.damage = 16;
-			item.knockBack = 0;
+			item.damage = 14;
+			item.knockBack = 2f;
 			item.ammo = AmmoID.Arrow;
 
 			item.ranged = true;
 			item.consumable = true;
 
-			item.shoot = ModContent.ProjectileType<Projectiles.Arrow.ShroomiteArrow>();
-			item.shootSpeed = 6f;
+			item.shoot = ModContent.ProjectileType<Projectiles.Arrow.SpectreArrow>();
+			item.shootSpeed = 5f;
 		}
 
 		public override void AddRecipes()
 		{
 			ModRecipe recipe = new ModRecipe(mod);
-			recipe.AddIngredient(ItemID.ShroomiteBar);
+			recipe.AddIngredient(ItemID.SpectreBar);
 			recipe.AddTile(TileID.MythrilAnvil);
 			recipe.SetResult(this, 50);
 			recipe.AddRecipe();
