+++ Items/Ammo/Arrow/TikiArrow.png
