+++ Items/Ammo/Bullet/MagicBullet.cs
@@ -1,17 +1,16 @@
 
-using SpiritMod.Projectiles.Arrow;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SpiritMod.Items.Ammo
+namespace SpiritMod.Items.Ammo.Bullet
 {
-	public class MarsBullet : ModItem
+	public class MagicBullet : ModItem
 	{
 		public override void SetStaticDefaults()
 		{
-			DisplayName.SetDefault("Electrified Bullet");
-			Tooltip.SetDefault("Zaps foes with powerful energy!");
+			DisplayName.SetDefault("Magic Bullet");
+			Tooltip.SetDefault("'What can't it do?'");
 		}
 
 
@@ -21,18 +20,17 @@ namespace SpiritMod.Items.Ammo
 			item.height = 16;
 			item.value = 1000;
 			item.rare = ItemRarityID.Red;
-			item.value = Item.buyPrice(0, 0, 3, 0);
-
+			item.value = Item.buyPrice(0, 0, 8, 0);
 			item.maxStack = 999;
 
-			item.damage = 15;
-			item.knockBack = 1.5f;
+			item.damage = 25;
+			item.knockBack = 2.5f;
 			item.ammo = AmmoID.Bullet;
 
 			item.ranged = true;
 			item.consumable = true;
 
-			item.shoot = ModContent.ProjectileType<MarsBulletProj>();
+			item.shoot = ModContent.ProjectileType<Projectiles.Bullet.MagicBullet>();
 			item.shootSpeed = 9f;
 
 		}
