+++ Items/Ammo/Bullet/SpectreBullet.cs
@@ -1,7 +1,7 @@
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SpiritMod.Items.Ammo
+namespace SpiritMod.Items.Ammo.Bullet
 {
 	public class SpectreBullet : ModItem
 	{
