+++ Items/Ammo/Bullet/SpectreBullet.png
