+++ NPCs/Town/Adventurer.cs
@@ -1,5 +1,5 @@
 using SpiritMod.Items.Accessory;
-using SpiritMod.Items.Ammo;
+using SpiritMod.Items.Ammo.Arrow;
 using SpiritMod.Items.Material;
 using SpiritMod.Items.Pins;
 using SpiritMod.Items.Placeable.Furniture;
